requirements.txt
cachetools==6.2.4
certifi==2025.11.12
charset-normalizer==3.4.4
google-api-core==2.28.1
google-api-python-client==2.193.0
google-auth==2.45.0
google-auth-httplib2==0.3.1
google-cloud-appengine-logging==1.9.0
google-cloud-audit-log==0.5.0
google-cloud-bigquery==3.41.0
google-cloud-core==2.5.0
google-cloud-firestore==2.22.0
google-cloud-logging==3.15.0
google-cloud-storage==3.10.1
google-crc32c==1.8.0
google-resumable-media==2.8.2
googleapis-common-protos==1.72.0
grpc-google-iam-v1==0.14.4
grpcio==1.76.0
grpcio-status==1.76.0
httplib2==0.31.2
idna==3.11
importlib_metadata==8.7.1
opentelemetry-api==1.40.0
packaging==26.0
proto-plus==1.27.0
protobuf==6.33.2
pyasn1==0.6.1
pyasn1_modules==0.4.2
pyparsing==3.3.2
python-dateutil==2.9.0.post0
python-dotenv==1.2.2
requests==2.32.5
rsa==4.9.1
six==1.17.0
typing_extensions==4.15.0
uritemplate==4.2.0
urllib3==2.6.2
zipp==3.23.0
